tetragon-metrics-docs: Move New function to a separate package
Introduce metricsmd package containing New function from
cmd/tetragon-metrics-docs. This is done so that this function can be imported
and reused by different commands.
